Revert "Merged tests for Binary Search Recursive and Non-Recursive"
